firmware-update: remove copy ota file to '/cache' partition from system level
Current Android SE policies does not allow System service to read files in
external storages (there is a 'nerverallow' rule for sdcard_type). This means
we cannot copy the ota update package at system level, so we need to move it
to the upper layer inside the API code.

This commit:
  * removes the start/end copy notify methods that are not required since the
    copy is done by the API code
  * defines common constant shared between API and service level

https://jira.digi.com/browse/ADK4A-1208

Signed-off-by: Tatiana Leon <tatiana.leon@digi.com>