ScanResult: convert batched scan results timestamp to ms
This is a temporal workaround: based on commit f2221848a720, the expected tsf
timestamp unit was us, but in our case (ath6kl driver), the timestamp filtering
is failing in upper layers so newer APs are not detected.
So, for workaround that, increase the timestamp and therefore newer APs will
not be filtered.

https://jira.digi.com/browse/ADK4A-1460

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>