PhoneWindowManager: cancel poweroff sequence after handling power key event
In ConnectCore8x platform, the MCA interprets the long power key
press as a shutdown action and starts the power-off sequence. If after
the configured guard time the device is not shutdown, the MCA watchdog
resets the MCA and the device automatically shuts down.

Android requirements are a bit different. Android needs to capture the
long power key press in order to display the global actions menu, but
that does not mean that the device should to shut-down at that moment.
For this reason, whenever the long power key process in Android finishes,
we need to cancel the MCA power-off sequence to avoid the module to
automatically shutdown. It is the user the one that decides what to do
(shut-down, reboot, etc.) after the long key press is captured and the
global actions menu is displayed.

In order to cancel the power-off sequence, we need to write a specific
string to one of the MCA sys file system entries as follows:

 echo "CANCEL PWROFF" > /sys/devices/platform/5a800000.i2c/i2c-0/0-0063/mca-cc8x-pwrkey/mca_cancel_pwroff

https://jira.digi.com/browse/ADK4A-1731

Signed-off-by: David Escalona <david.escalona@digi.com>
(cherry picked from commit 75e9c2f728e9372664bb92d22aaf967146fcf3a5)
Signed-off-by: Javier Viguera <javier.viguera@digi.com>