power_manager_service: disable/enable WiFi/BT interfaces on suspend/wakeup
This commit adds the support to disable and enable the Wi-Fi and Bluetooth interfaces
when the system goes to suspend based on an Android boot read only property. This allows
the device to switch off the 3.3V regulator on suspend mode without breaking these two
interfaces when the device wakes up.

https://jira.digi.com/browse/ADK4A-1767

Signed-off-by: David Escalona <david.escalona@digi.com>